# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Микросервис для автоматического назначения ревьюеров на Pull Request'ы",
        "title": "PR Reviewer Assignment Service",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "shkeera@mail.ru"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pullRequest/create": {
            "post": {
                "description": "Создает новый PR и автоматически назначает ревьюеров",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull-requests"
                ],
                "summary": "Создание Pull Request",
                "parameters": [
                    {
                        "description": "Данные Pull Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePRRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный PR",
                        "schema": {
                            "$ref": "#/definitions/handler.PRResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "PR уже существует",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "description": "Помечает PR как мердженный",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull-requests"
                ],
                "summary": "Merge Pull Request",
                "parameters": [
                    {
                        "description": "ID Pull Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MergePRRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный PR",
                        "schema": {
                            "$ref": "#/definitions/handler.PRResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "description": "Заменяет ревьюера на PR на другого активного участника команды",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull-requests"
                ],
                "summary": "Замена ревьюера",
                "parameters": [
                    {
                        "description": "Данные для замены ревьюера",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ReassignReviewerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат замены",
                        "schema": {
                            "$ref": "#/definitions/handler.ReassignReviewerResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR или ревьюер не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Ошибка замены ревьюера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/getReview": {
            "get": {
                "description": "Возвращает список PR, назначенных на пользователя для ревью",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение назначенных PR пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список PR пользователя",
                        "schema": {
                            "$ref": "#/definitions/handler.UserPRsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/setIsActive": {
            "post": {
                "description": "Активирует или деактивирует пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Установка активности пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SetUserActiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный пользователь",
                        "schema": {
                            "$ref": "#/definitions/handler.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.CreatePRRequest": {
            "type": "object",
            "required": [
                "author_id",
                "pull_request_id",
                "pull_request_name"
            ],
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "user-456"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-123"
                },
                "pull_request_name": {
                    "type": "string",
                    "example": "Fix login issue"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "handler.MergePRRequest": {
            "type": "object",
            "required": [
                "pull_request_id"
            ],
            "properties": {
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-123"
                }
            }
        },
        "handler.PRResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/models.PullRequest"
                }
            }
        },
        "handler.ReassignReviewerRequest": {
            "type": "object",
            "required": [
                "old_reviewer_id",
                "pull_request_id"
            ],
            "properties": {
                "old_reviewer_id": {
                    "type": "string",
                    "example": "user-789"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-123"
                }
            }
        },
        "handler.ReassignReviewerResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/models.PullRequest"
                },
                "replaced_by": {
                    "type": "string"
                }
            }
        },
        "handler.SetUserActiveRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "user_id": {
                    "type": "string",
                    "example": "user-123"
                }
            }
        },
...
        "handler.UserPRsResponse": {
            "type": "object",
            "properties": {
                "pull_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PullRequestShort"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handler.UserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.PullRequest": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "mergedAt": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.PullRequestShort": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "team_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}